use double instead of long double

+++ src/System/Tensor2_3D.cpp
@@ -599,7 +599,7 @@ Tensor2_3D Ten2::gen_zero()
 Tensor2_3D Ten2::exp( const Tensor2_3D &input )
 {
   double nn = 0.0;
-  long double nn_fac = 1.0;
+  double nn_fac = 1.0;
 
   Tensor2_3D input_pow = Ten2::gen_id();
   Tensor2_3D input_exp = Ten2::gen_id();
